% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_backup_googledrive.R
\name{criar_diretorio_googledrive}
\alias{criar_diretorio_googledrive}
\title{Função que executa o backup dos dados}
\usage{
criar_diretorio_googledrive()
}
\description{
Função que executa o backup dos dados
}
