% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_criar_tabelas_bd.R
\name{criar_tabelas_bd}
\alias{criar_tabelas_bd}
\title{Função que cria as Tabelas do Banco de Dados}
\usage{
criar_tabelas_bd(sgbd = "sqlite")
}
\arguments{
\item{sgbd}{Define o Sistema de Banco de Dados a ser utilizado. Por padrão, é definido como sqlite}
}
\description{
Função que cria as Tabelas do Banco de Dados
}
