% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_criar_tb_dcalendario.R
\name{criar_tb_dcalendario}
\alias{criar_tb_dcalendario}
\title{Função que gera a tabela com a relação de meses e ano para o Web Scraping;}
\usage{
criar_tb_dcalendario(anos, sgbd = "sqlite")
}
\description{
Função que gera a tabela com a relação de meses e ano para o Web Scraping;
}
