% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_criar_tb_dmunicipios.R
\name{criar_tb_dmunicipios}
\alias{criar_tb_dmunicipios}
\title{Função que gera a tabela "tabela_tcm_dmunicipios"}
\usage{
criar_tb_dmunicipios()
}
\description{
Função que gera a tabela "tabela_tcm_dmunicipios"
}
