% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_criar_tb_requisicoes.R
\name{criar_tb_requisicoes_pessoal}
\alias{criar_tb_requisicoes_pessoal}
\title{Função que gera a Tabela de Requisições "tabela_paginas_links" no Banco de Dados}
\usage{
criar_tb_requisicoes_pessoal(sgbd = "sqlite")
}
\arguments{
\item{sgbd}{Define o Sistema de Banco de Dados a ser utilizado. Por padrão, é definido como sqlite}
}
\description{
Função que gera a Tabela de Requisições "tabela_paginas_links" no Banco de Dados
}
