% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_backup_dropbox.R
\name{executar_backup_bd_dropbox}
\alias{executar_backup_bd_dropbox}
\title{Função que executa o backup}
\usage{
executar_backup_bd_dropbox(backup = "NAO")
}
\description{
Função que executa o backup do Banco de Dados
e dos arquivos CSV para o DropBox
}
