% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_backup_dropbox.R
\name{executar_backup_csv_dropbox}
\alias{executar_backup_csv_dropbox}
\title{Função que executa o backup dos dados}
\usage{
executar_backup_csv_dropbox(backup = "NAO")
}
\description{
Função que executa o backup dos dados
}
