% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_backup_googledrive.R
\name{executar_backup_csv_googledrive}
\alias{executar_backup_csv_googledrive}
\title{Função que executa o backup dos dados}
\usage{
executar_backup_csv_googledrive(backup = "NAO")
}
\arguments{
\item{backup}{É definido como "NAO" como padrão. Mas pode ser marcado como "SIM"
para realizar o Backup do Banco de Dados e dos arquivos CSV.}
}
\description{
Função que executa o backup dos dados
}
