% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_exportar_dropbox.R
\name{exportar_csv_unico_dropbox}
\alias{exportar_csv_unico_dropbox}
\title{Função que exporta os arquivos CSV único para o DropBox, com o objetivo
conectar com o Power BI}
\usage{
exportar_csv_unico_dropbox(backup = "NAO")
}
\description{
Função que exporta os arquivos CSV único para o DropBox, com o objetivo
conectar com o Power BI
}
