% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{limpar_nome}
\alias{limpar_nome}
\title{Função para limpar nomes de caracteres especiais}
\usage{
limpar_nome(x)
}
\arguments{
\item{x}{Limpa o nome dos entes municipais para retirar caracteres especiais}
}
\description{
Função para limpar nomes de caracteres especiais
}
