% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcmbapessoal-package.R
\docType{package}
\name{tcmbapessoal-package}
\alias{tcmbapessoal}
\alias{tcmbapessoal-package}
\title{tcmbapessoal: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
