% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valor_monetario}
\alias{valor_monetario}
\title{Função utilizadas no Web Scraping}
\usage{
valor_monetario(x)
}
\arguments{
\item{x}{Limpa os valores monetários de caracteres especiais}
}
\description{
Função utilizadas no Web Scraping
}
