% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valor_monetario2}
\alias{valor_monetario2}
\title{Função utilizadas no Web Scraping}
\usage{
valor_monetario2(x)
}
\arguments{
\item{x}{string}
}
\description{
Transforma valores Ex: c("R$ 10.000,89", "R$ 123,75") em 10000.89 e 123.75
}
