% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_criar_tabelas_bd.R
\name{criar_tabelas_bd}
\alias{criar_tabelas_bd}
\title{Função que cria as tabelas do Banco de Dados no SQLite}
\usage{
criar_tabelas_bd()
}
\description{
Função que cria as tabelas do Banco de Dados no SQLite
}
