% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_criar_tabelas_bd.R
\name{criar_tabelas_bd}
\alias{criar_tabelas_bd}
\title{Função que cria as Tabelas do Banco de Dados}
\usage{
criar_tabelas_bd(sgbd = "sqlite")
}
\description{
Função que cria as Tabelas do Banco de Dados
}
