% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_criar_tb_requisicoes.R
\name{criar_tb_requisicoes_pessoal}
\alias{criar_tb_requisicoes_pessoal}
\title{Função que gera a Tabela de Requisições "tabela_paginas_links"}
\usage{
criar_tb_requisicoes_pessoal()
}
\description{
Função que gera a Tabela de Requisições "tabela_paginas_links"
}
