% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_backup_arquivos.R
\name{executar_backup_arquivos}
\alias{executar_backup_arquivos}
\title{Função que executa o backup}
\usage{
executar_backup_arquivos(backup = "NAO")
}
\description{
Função que executa o backup do Banco de Dados, dos arquivos HTML,
e dos arquivos CSVs.
}
