% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_executar_backup_googledrive.R
\name{executar_backup_csv_googledrive}
\alias{executar_backup_csv_googledrive}
\title{Função que executa o backup dos dados}
\usage{
executar_backup_csv_googledrive(backup = "NAO")
}
\description{
Função que executa o backup dos dados
}
