% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_scraping_html_folhapessoal.R
\name{executar_scraping_html_folhapessoal}
\alias{executar_scraping_html_folhapessoal}
\title{Função que executa o Web Scraping}
\usage{
executar_scraping_html_folhapessoal(repetir = "NAO", sgbd = "sqlite")
}
\arguments{
\item{repetir}{É definido como "NAO" por padrão, mas pode ser atribuído SIM
para realizar o Web Scraping de páginas que retornaram erro ou que ainda não foram informados pelos
municípios ao TCM-Ba}

\item{sgbd}{Define o Sistema de Banco de Dados a ser utilizado. Por padrão, é definido como sqlite}
}
\description{
Executa o Web Scraping das Páginas que tem a Folha de pessoal dos Municípios no site do TCM-ba
}
