% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_scraping_html_folhapessoal.R
\name{executar_scraping_html_folhapessoal}
\alias{executar_scraping_html_folhapessoal}
\title{Função que executa as requisições do Web Scraping}
\usage{
executar_scraping_html_folhapessoal(repetir = "SIM", sgbd = "sqlite")
}
\arguments{
\item{repetir}{É definido "SIM" como padrão. Mas pode ser marcado como "NAO",
caso não deseje repetir as consulta do Web Scraping que falharam ou que não foram
identificadas resposta do ente municipal no dia de execução do Web Scraping}

\item{sgbd}{Define o Sistema de Banco de Dados a ser utilizado. Por padrão, é definido como sqlite}
}
\description{
Realiza as requisições das Páginas (HTML)
que tem a Folha de pessoal dos Municípios no site do TCM-ba.
}
