% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_scraping_html_folhapessoal.R
\name{executar_scraping_html_folhapessoal}
\alias{executar_scraping_html_folhapessoal}
\title{Função que executa o Web Scraping das Páginas de despesas do TCM-ba}
\usage{
executar_scraping_html_folhapessoal(repetir = "NAO")
}
\description{
Função que executa o Web Scraping das Páginas de despesas do TCM-ba
}
