% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_executar_web_scraping.R
\name{executar_web_scraping}
\alias{executar_web_scraping}
\title{Função que executa o Web Scraping}
\usage{
executar_web_scraping(ano, nome_scraping, repetir, backup)
}
\description{
Essa função foi desenvolvida ...
}
