% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_executar_web_scraping.R
\name{executar_web_scraping}
\alias{executar_web_scraping}
\title{Função que executa o Web Scraping}
\usage{
executar_web_scraping(ano, nome_scraping, bd, repetir, backup)
}
\arguments{
\item{ano}{Exercíico (ano) que ser pretende obter}

\item{nome_scraping}{Nome do Diretório que será criado para alocar o Web Scraping}

\item{bd}{Define qual é o Banco de Dados a ser utilizado.
Por padrão, é definido o SQLite}

\item{repetir}{É definido "NAO" como padrão. Mas pode ser marcado como "SIM",
caso deseje repetir as consulta do Web Scraping que falharam ou que não foram
identificadas resposta do ente municipal}

\item{backup}{É definido como "NAO" como padrão. Mas pode ser marcado como "SIM"
para realizar o Backup do Banco de Dados e dos arquivos CSV para o Google Drive.
É preciso configurar o token Google Drive antes de executar}
}
\description{
Essa função foi desenvolvida para executar todas as etapas do Web Scraping
}
