% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_exportar_dropbox.R
\name{exportar_csv_dropbox}
\alias{exportar_csv_dropbox}
\title{Função que exporta os arquivos CSV para o DropBox, com o objetivo
conectar com o Power BI}
\usage{
exportar_csv_dropbox(backup = "NAO")
}
\arguments{
\item{backup}{É definido como "NAO" como padrão. Mas pode ser marcado como "SIM"
para realizar o Backup do Banco de Dados e dos arquivos CSV.}
}
\description{
Função que exporta os arquivos CSV para o DropBox, com o objetivo
conectar com o Power BI
}
