% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{limpar_nome}
\alias{limpar_nome}
\title{Função para limpar nomes de caracteres especiais}
\usage{
limpar_nome(x)
}
\arguments{
\item{x}{string}
}
\description{
Limpa o nome dos entes municipais para retirar caracteres especiais
}
\details{
#' @title Função utilizadas no Web Scraping
#'
#' @description Transforma valores Ex: c("R$ 10.000,89", "R$ 123,75") em 10000.89 e 123.75
#'
#' @param x string
#'
#' @export

valor_monetario2 <- function(x) {
x %>%
stringr::str_remove("R\$") %>%
stringr::str_remove(".") %>%
stringr::str_replace_all(",", "\.") %>%
as.character()
}
}
