% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcmbapessoal-package.R
\docType{package}
\name{tcmbapessoal-package}
\alias{tcmbapessoal}
\alias{tcmbapessoal-package}
\title{tcmbapessoal: Web Scraping das Folhas de Pessoas dos Municipios do Estado da Bahia}
\description{
Web Scraping das Folhas de Pessoas dos Municipios do Estado da Bahia
                pelo site do TCM-Ba
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/georgevbsantiago/tcmbapessoal}
}

}
\author{
\strong{Maintainer}: George Santiago \email{georgevbsantiago@hotmail.com}

}
\keyword{internal}
