% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcmbapessoal-package.R
\docType{package}
\name{tcmbapessoal-package}
\alias{tcmbapessoal}
\alias{tcmbapessoal-package}
\title{tcmbapessoal: Web Scraping das Folhas de Pessoas dos Municipios do Estado da Bahia}
\description{
Web Scraping desenvolvido pelo Observatório Social de Santo Antonio de Jesus (Brasil/Bahia).
    Tem como objetivo obter e tratar os dados das Folhas de Pessoas dos Municipios do Estado da Bahia
    disponibilizados no site (na estrutura HTML) do TCM-Ba.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/georgevbsantiago/tcmbapessoal}
}

}
\author{
\strong{Maintainer}: George Santiago \email{georgevbsantiago@hotmail.com}

}
\keyword{internal}
